LPS-109844 Move DDMFormValuesUtil to util.v1_0
