+++ modules/apps/headless/headless-delivery/headless-delivery-api/src/main/java/com/liferay/headless/delivery/util/v1_0/DDMFormValuesUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.headless.delivery.dynamic.data.mapping;
+package com.liferay.headless.delivery.util.v1_0;
 
 import com.liferay.document.library.kernel.service.DLAppService;
 import com.liferay.dynamic.data.mapping.form.field.type.constants.DDMFormFieldTypeConstants;
@@ -27,7 +27,6 @@ import com.liferay.dynamic.data.mapping.model.Value;
 import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
 import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
 import com.liferay.headless.delivery.dto.v1_0.ContentField;
-import com.liferay.headless.delivery.util.v1_0.DDMValueUtil;
 import com.liferay.journal.service.JournalArticleService;
 import com.liferay.petra.function.UnsafeFunction;
 import com.liferay.portal.kernel.service.LayoutLocalService;
