+++ modules/apps/headless/headless-delivery/headless-delivery-impl/src/main/java/com/liferay/headless/delivery/internal/resource/v1_0/StructuredContentResourceImpl.java
@@ -37,7 +37,6 @@ import com.liferay.headless.delivery.dto.v1_0.ContentField;
 import com.liferay.headless.delivery.dto.v1_0.Rating;
 import com.liferay.headless.delivery.dto.v1_0.StructuredContent;
 import com.liferay.headless.delivery.dynamic.data.mapping.DDMFormFieldUtil;
-import com.liferay.headless.delivery.dynamic.data.mapping.DDMFormValuesUtil;
 import com.liferay.headless.delivery.internal.dto.v1_0.converter.StructuredContentDTOConverter;
 import com.liferay.headless.delivery.internal.dto.v1_0.util.DisplayPageRendererUtil;
 import com.liferay.headless.delivery.internal.dto.v1_0.util.EntityFieldsUtil;
@@ -50,6 +49,7 @@ import com.liferay.headless.delivery.search.aggregation.AggregationUtil;
 import com.liferay.headless.delivery.search.filter.FilterUtil;
 import com.liferay.headless.delivery.search.sort.SortUtil;
 import com.liferay.headless.delivery.util.v1_0.CustomFieldsUtil;
+import com.liferay.headless.delivery.util.v1_0.DDMFormValuesUtil;
 import com.liferay.headless.delivery.util.v1_0.DDMValueUtil;
 import com.liferay.headless.delivery.util.v1_0.StructuredContentUtil;
 import com.liferay.info.item.InfoItemServiceTracker;
